ajustes sicronizacion

+++ GeotimeConnectWebApi/Models/cMarcaResumen.cs
@@ -8,7 +8,7 @@ namespace GeoTimeConnectWebApi.Models
         public String? IdNumero { get; set; }
         public int IdConcepto { get; set; } 
         public String? NominaEq { get; set; }
-        public decimal Cantidad { get; set; }
+        public double Cantidad { get; set; }
         public decimal Monto { get; set; }
         public String? IdCCosto { get; set; }
         public String? Proyecto { get; set; }
